* Refactored code into different packages * Changed getText() methods to getString to eliminate needless casting * Moved checkWin and checkDraw methods to new class GameLogic * Improved layouts to have buttons which use styles * Reduced the min SDK version to sensible level * Moved strings to strings.xml
