+++ app/src/main/java/com/codingnomads/andy/tictactoe/players/Player.java
@@ -1,4 +1,4 @@
-package com.codingnomads.andy.tictactoe;
+package com.codingnomads.andy.tictactoe.players;
 
 public enum Player {
     PLAYER_ONE,
