+++ app/src/main/res/values/strings.xml
@@ -11,4 +11,5 @@
     <string name="easy">Easy</string>
     <string name="medium">Medium</string>
     <string name="hard">Impossible</string>
+    <string name="new_game">New Game</string>
 </resources>
